refactor: mouse moved handler refactored
