+++ src/input_device_monitor/event_handler/concrete/mouse_event_handler.rs
@@ -2,7 +2,8 @@ use nannou::event::WindowEvent;
 use std::sync::{Arc, Mutex};
 
 use crate::input_device_monitor::{
-    event_handler::AEventHandler, my_event::serializable_clone::update_last_mouse_point, sender::IEventSender
+    event_handler::{mouse_moved::update_last_mouse_point, AEventHandler},
+    sender::IEventSender,
 };
 
 pub struct MouseEventHandler {
