+++ src/input_device_monitor/event_handler/mod.rs
@@ -1,4 +1,5 @@
 pub mod event_handler;
+pub mod mouse_moved;
 pub mod concrete;
 
 pub use event_handler::{AEventHandler, EventHandler};
