+++ src/input_device_monitor/event_handler/mouse_moved/mod.rs
@@ -1,0 +1,3 @@
+pub mod mouse_movev_handler;
+
+pub use mouse_movev_handler::{get_last_mouse_point, update_last_mouse_point};
