+++ src/input_device_monitor/my_event/serializable_clone/mod.rs
@@ -5,5 +5,5 @@ pub mod window;
 
 pub use key::MyKey;
 pub use mouse_button::MyMouseButton;
-pub use vec2::{MyVec2, update_last_mouse_point, get_last_mouse_point};
+pub use vec2::MyVec2;
 pub use window::MyWindowEvent;
